backend/requirements.txt
alembic==1.14.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
catboost==1.2.7
category-encoders==2.6.4
cloudpickle==3.1.0
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
eli5==0.13.0
executing==2.1.0
fonttools==4.54.1
graphviz==0.20.3
imageio==2.36.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
lightgbm==4.5.0
lime==0.2.0.1
llvmlite==0.43.0
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
optuna==4.1.0
optuna-integration==4.1.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
phik==0.12.4
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyMySQL==1.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
shap==0.46.0
six==1.16.0
slicer==0.0.8
SQLAlchemy==2.0.36
sqlparse==0.5.0
stack-data==0.6.3
statsmodels==0.14.4
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
wcwidth==0.2.13
xgboost==2.1.2
